請看詳細
2019/03/01 更新:
1.新增BOSS大量掉寶(西格諾斯、阿卡)
2.翻譯專業技術村NPC
3.修復疲勞值
4.設定營養劑使用次數 - 每日每個帳號3次
5.設定技術村採集地圖 - 每張圖每日每個帳號3次
6.修復特殊道具箱子無法開啟問題
7.現在FB未成年者回復也可取得獎勵(但粉專無法私訊告知)
8.修復賢者之石等稀有素材無法取得問題
9.新增靈魂之石與第六天靈魂石進入封測道具

2019/02/28更新:
1.轉蛋大更新(已應用)
2.修復部分點裝滑鼠移過去會斷線問題
3.修復部分椅子使用後會斷線問題
4.修復衝方塊楓幣費用消耗錯誤問題
5.修復破攻強化腳本錯誤問題
6.修復@sell指令 賣道具時獲得楓幣計算錯誤問題
7.創建公會與腳色限制部分不雅字眼
P.S.若之後出現不雅名字，GM有權利更改為我愛小喵谷XXXX
8.修復暗殺者、神偷，影子無效果
9.修正弓箭手職業招喚招喚獸後會無傷害問題
10.修復機甲戰神有時候無法轉換回一般狀態問題

+++ src/scripting/ReactorActionManager.java
@@ -199,7 +199,7 @@ public class ReactorActionManager extends AbstractPlayerInteraction {
     }
 
     public void cancelHarvest(boolean succ) {
-        getPlayer().setFatigue((byte) (getPlayer().getFatigue() + 1));
+        getPlayer().setFatigue((getPlayer().getFatigue() + 1));
         getPlayer().getMap().broadcastMessage(getPlayer(), CField.showHarvesting(getPlayer().getId(), 0), false);
         getPlayer().getMap().broadcastMessage(CField.harvestResult(getPlayer().getId(), succ));
     }
